Fix and optimize queries and generate ordering model

+++ db/migrate/20220621120505_create_cart_items.rb
@@ -1,7 +1,7 @@
 class CreateCartItems < ActiveRecord::Migration[7.0]
   def change
     create_table :cart_items do |t|
-      t.numeric :quantity, null: false
+      t.integer :quantity, null: false
       t.belongs_to :product, foreign_key: true, index: true
       t.belongs_to :cart, foreign_key: true, index: true, type: :uuid
